ToConsiderImplementLogic/mcp-adapter-main/examples/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
sqlite3>=2.6.0
docker>=7.0.0
google-generativeai>=0.3.0
